# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Pet-project GO-Friend-Sphere REST API",
        "title": "GO-Friend-Sphere",
        "contact": {
            "email": "polozenko.kirill.job@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/friendships": {
            "post": {
                "description": "create a new friendship",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friendships"
                ],
                "summary": "Create Friendship",
                "parameters": [
                    {
                        "description": "Friendship object to be created",
                        "name": "friendship",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Friendship"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Friendship"
                        }
                    }
                }
            }
        },
        "/friendships/user/{userId}": {
            "get": {
                "description": "get a list of friendships by user ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friendships"
                ],
                "summary": "Get Friendships by User ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Friendship"
                            }
                        }
                    }
                }
            }
        },
        "/friendships/{friendshipId}": {
            "get": {
                "description": "get a friendship by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friendships"
                ],
                "summary": "Get Friendship by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Friendship ID",
                        "name": "friendshipId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Friendship"
                        }
                    }
                }
            },
            "put": {
                "description": "update a friendship's status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friendships"
                ],
                "summary": "Update Friendship",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Friendship ID",
                        "name": "friendshipId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Friendship"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete a friendship",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friendships"
                ],
                "summary": "Delete Friendship",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Friendship ID",
                        "name": "friendshipId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/messages": {
            "post": {
                "description": "create a new message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Create Message",
                "parameters": [
                    {
                        "description": "Message object to be created",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/messages/user/{userId}": {
            "get": {
                "description": "get a list of messages by user ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Get Messages by User ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Message"
                            }
                        }
                    }
                }
            }
        },
        "/messages/{messageId}": {
            "get": {
                "description": "get a message by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Get Message by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Message ID",
                        "name": "messageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            },
            "put": {
                "description": "update a message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Update Message",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Message ID",
                        "name": "messageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete a message",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Delete Message",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Message ID",
                        "name": "messageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Friendship": {
            "type": "object",
            "required": [
                "friend_id",
                "status",
                "user_id"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "friend_id": {
                    "type": "integer"
                },
                "friendship_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string",
                    "maxLength": 24
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.Message": {
            "type": "object",
            "required": [
                "message",
                "recipient_id",
                "sender_id"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "maxLength": 512
                },
                "message_id": {
                    "type": "integer"
                },
                "read_at": {
                    "type": "string"
                },
                "recipient_id": {
                    "type": "integer"
                },
                "sender_id": {
                    "type": "integer"
                },
                "sent_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}